Add functionality to accept and reject buttons in driver view.

+++ app/src/main/java/com/jbseppanen/birthride/Models.kt
@@ -125,5 +125,5 @@ data class RideRequestData(
     val name: String,
     val phone: String,
     val price: Int,
-    val ride_id: Int
+    val ride_id: Long
 )
